Finalize fastols api
