+++ notebooks/dev/FastOLS.py
@@ -28,14 +28,14 @@ def _():
     import logging
 
     configure_logging(level=logging.DEBUG)
-    return CamlSyntheticDataGenerator, FastOLS, np
+    return CamlSyntheticDataGenerator, FastOLS
 
 
 @app.cell
 def _(CamlSyntheticDataGenerator):
     data = CamlSyntheticDataGenerator(
         n_obs=100_000,
-        n_cont_outcomes=2,
+        n_cont_outcomes=1,
         n_binary_outcomes=0,
         n_cont_treatments=0,
         n_binary_treatments=1,
@@ -44,8 +44,8 @@ def _(CamlSyntheticDataGenerator):
         n_cont_modifiers=1,
         n_binary_modifiers=2,
         n_discrete_modifiers=1,
-        stddev_outcome_noise=0.1,
-        stddev_treatment_noise=0.1,
+        stddev_outcome_noise=10,
+        stddev_treatment_noise=10,
         causal_model_functional_form="linear",
         seed=44,
     )
@@ -55,7 +55,7 @@ def _(CamlSyntheticDataGenerator):
 @app.cell
 def _(data):
     df = data.df
-    # df["cates"] = data.cates
+    df["cates"] = data.cates
     df
     return (df,)
 
@@ -74,7 +74,7 @@ def _(FastOLS, df):
         G=[c for c in df.columns if "X" in c and ("bin" in c or "dis" in c)],
         X=[c for c in df.columns if "X" in c and "cont" in c],
         W=[c for c in df.columns if "W" in c],
-        engine="cpu",
+        engine="gpu",
         discrete_treatment=True,
     )
     return (fu,)
@@ -82,59 +82,30 @@ def _(FastOLS, df):
 
 @app.cell
 def _(df, fu):
-    fu.fit(data=df, n_jobs=-1, estimate_effects=True, robust_vcv=True)
+    fu.fit(df, n_jobs=-1, estimate_effects=True, robust_vcv=False)
+
+    fu.results
     return
 
 
 @app.cell
-def _():
-    import jax.numpy as jnp
-    return (jnp,)
+def _(df, fu):
+    cates = fu.estimate_cate(df, return_results_dict=True)
 
-
-@app.cell
-def _(dmatrix, fu, jnp):
-    dbar = jnp.mean(dmatrix, axis=0)
-
-    dbar @ fu.results['vcv'] @ dbar
+    cates
     return
 
 
 @app.cell
-def _(np):
-    # Create example arrays
-    a = np.random.rand(15)       # 1D array of shape (15)
-    B = np.random.rand(2, 15, 15) # 3D tensor of shape (2, 15, 15)
-    c = np.random.rand(15)       # 1D array of shape (15)
-
-    print(a.shape)
-    print(B.shape)
-    print(c.shape)
-    # # To multiply these, you need to decide how to handle the first dimension of B
-    # # For example, you could compute this for each slice of B:
-    results = []
-    for i in range(2):
-        # For each slice B[i], compute a @ B[i] @ c
-        temp = a @ B[i] @ a
-        results.append(temp)
-    return B, a, results
-
-
-@app.cell
-def _(B, a):
-    a @ B @ a
+def _(df, fu):
+    cate_predictions = fu.predict(df)
+    cate_predictions
     return
 
 
 @app.cell
-def _(results):
-    results
-    return
-
-
-@app.cell
-def _(fu):
-    fu.results['vcv']
+def _(data):
+    data.cates
     return
 
 
@@ -165,18 +136,13 @@ def _(df, fu):
         "X2_binary == 0 & X4_discrete == 3 & X1_continuous < 5"
     ).copy()
 
-    fu.prettify_treatment_effects(effects=fu.estimate_ate(data=df2))
+    fu.prettify_treatment_effects(effects=fu.estimate_ate(df2, return_results_dict=True))
     return (df2,)
 
 
 @app.cell
 def _(df2):
     df2['cates'].mean()
-    return
-
-
-@app.cell
-def _():
     return
